feat: Update dependencies and add new packages
This commit updates the frontend/package.json file to include the "youtube-vue3" package as a new dependency with version "^0.1.15". Additionally, it updates the backend/package.json file to include the "@csstools/postcss-minify" package with version "^1.1.3" and the "@google/generative-ai" package with version "^0.11.1". These updates and additions ensure that the project has the latest versions of the required packages and introduces new functionality related to YouTube integration and CSS minification.