feat: Add getStream method to Channel model
This commit adds a new method, getStream, to the Channel model in order to retrieve the current stream associated with the channel. The method makes use of the Twitch API to fetch the stream information based on the Twitch ID of the channel. This enhancement improves the functionality of the application by providing access to the current stream details for a channel.